refactor(checkbox,radio): dedupe focus styles and item input observer
