+++ core/src/components/checkbox/checkbox.ios.scss
@@ -31,26 +31,3 @@
 :host(.checkbox-disabled) {
   opacity: $checkbox-ios-disabled-opacity;
 }
-
-// iOS Checkbox: Keyboard Focus
-// -----------------------------------------
-
-:host(.ion-focused) .native-wrapper::after {
-  @include border-radius(50%);
-
-  display: block;
-  position: absolute;
-
-  width: 36px;
-  height: 36px;
-
-  transform: translate(-50%, -50%);
-
-  background: $checkbox-ios-background-color-focused;
-
-  content: "";
-  opacity: 0.2;
-
-  inset-block-start: 50%;
-  inset-inline-start: 50%;
-}
