+++ core/src/components/checkbox/checkbox.ios.vars.scss
@@ -28,6 +28,3 @@ $checkbox-ios-disabled-opacity:           $form-control-ios-disabled-opacity;
 
 /// @prop - Checkmark width of the checkbox icon
 $checkbox-ios-icon-checkmark-width:       1.5px;
-
-/// @prop - Background color of focus indicator for checkbox when focused
-$checkbox-ios-background-color-focused:   ion-color(primary, tint);
