+++ core/src/components/checkbox/checkbox.md.scss
@@ -51,26 +51,3 @@
 :host(.checkbox-disabled) .native-wrapper {
   opacity: $checkbox-md-icon-disabled-opacity;
 }
-
-// Material Design Checkbox: Keyboard Focus
-// -----------------------------------------
-
-:host(.ion-focused) .native-wrapper::after {
-  @include border-radius(50%);
-
-  display: block;
-  position: absolute;
-
-  width: 36px;
-  height: 36px;
-
-  transform: translate(-50%, -50%);
-
-  background: $checkbox-md-background-color-focused;
-
-  content: "";
-  opacity: 0.2;
-
-  inset-block-start: 50%;
-  inset-inline-start: 50%;
-}
