+++ core/src/components/checkbox/checkbox.tsx
@@ -1,6 +1,6 @@
 import type { ComponentInterface, EventEmitter } from '@stencil/core';
 import { Build, Component, Element, Event, Host, Method, Prop, State, forceUpdate, h } from '@stencil/core';
-import { checkInvalidState } from '@utils/forms';
+import { checkInvalidState, createItemMultipleInputsObserver } from '@utils/forms';
 import type { Attributes } from '@utils/helpers';
 import { inheritAriaAttributes, renderHiddenInput } from '@utils/helpers';
 import { createColorClasses, hostContext } from '@utils/theme';
@@ -201,21 +201,9 @@ export class Checkbox implements ComponentInterface {
     this.isInvalid = checkInvalidState(el);
     this.hasLabelContent = this.el.textContent !== '';
 
-    // The item toggles `item-multiple-inputs` after this control renders and as
-    // inputs are added or removed. Re-render when it flips so the focus
+    // Re-render when the item flips `item-multiple-inputs` so the focus
     // indicator stays in sync.
-    const item = el.closest('ion-item');
-    if (item && Build.isBrowser && typeof MutationObserver !== 'undefined') {
-      let wasMultipleInputs = item.classList.contains('item-multiple-inputs');
-      this.itemFocusObserver = new MutationObserver(() => {
-        const isMultipleInputs = item.classList.contains('item-multiple-inputs');
-        if (isMultipleInputs !== wasMultipleInputs) {
-          wasMultipleInputs = isMultipleInputs;
-          forceUpdate(this);
-        }
-      });
-      this.itemFocusObserver.observe(item, { attributes: true, attributeFilter: ['class'] });
-    }
+    this.itemFocusObserver = createItemMultipleInputsObserver(el, () => forceUpdate(this));
   }
 
   componentWillLoad() {
