+++ core/src/components/checkbox/test/basic/checkbox.e2e.ts
@@ -288,8 +288,7 @@ configs({ modes: ['ios'], directions: ['ltr'] }).forEach(({ title, config }) => 
 });
 
 /**
- * The focus indicator UI differs between iOS and MD, so these visual tests run
- * in both modes. Direction does not affect the indicator, so only LTR is run.
+ * This behavior does not vary across directions
  */
 configs({ directions: ['ltr'] }).forEach(({ title, screenshot, config }) => {
   test.describe(title('checkbox: focus visual'), () => {
