+++ core/src/components/radio/test/basic/radio.e2e.ts
@@ -2,8 +2,7 @@ import { expect } from '@playwright/test';
 import { configs, test } from '@utils/test/playwright';
 
 /**
- * The focus indicator UI differs between iOS and MD, so these visual tests run
- * in both modes. Direction does not affect the indicator, so only LTR is run.
+ * This behavior does not vary across directions
  */
 configs({ directions: ['ltr'] }).forEach(({ title, screenshot, config }) => {
   test.describe(title('radio: focus visual'), () => {
