+++ core/src/components/radio/test/item/radio.e2e.ts
@@ -101,8 +101,7 @@ configs({ directions: ['ltr'] }).forEach(({ title, screenshot, config }) => {
 });
 
 /**
- * The focus indicator UI differs between iOS and MD, so these visual tests run
- * in both modes. Direction does not affect the indicator, so only LTR is run.
+ * This behavior does not vary across directions
  */
 configs({ directions: ['ltr'] }).forEach(({ title, screenshot, config }) => {
   test.describe(title('radio: focus in item'), () => {
