+++ core/src/utils/forms/index.ts
@@ -1,3 +1,4 @@
 export * from './notch-controller';
 export * from './compare-with-utils';
+export * from './item-multiple-inputs';
 export * from './validity';
